+++ Sources/VCore/Extensions/Core Frameworks/NSLayoutConstraint.Storing.swift
@@ -5,9 +5,13 @@
 //  Created by Vakhtang Kontridze on 9/13/21.
 //
 
-#if canImport(UIKit) && !os(watchOS)
+#if !os(watchOS)
 
+#if canImport(UIKit)
 import UIKit
+#elseif canImport(AppKit)
+import AppKit
+#endif
 
 // MARK: - Storing Layout Constraint
 extension NSLayoutConstraint {